% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_hom.R
\name{F_hom}
\alias{F_hom}
\title{F_hom: calculates FHOM and reports basic statistics}
\usage{
F_hom(het_path)
}
\arguments{
\item{het_path}{A character string indicating the path of the .het PLINK output}
}
\value{
A list, containing:
\item{FHOM_values}{A data frame with FID, IID, and Fhom values}
\item{Fhom_summary_table}{A data frame with overall Fhom summary statistics (Maximum, Minimum, Mean, SD)}
\item{by_population_Fhom_summary}{A data frame with per population Fhom summary statistics (Maximum, Minimum, Mean, SD)}
}
\description{
F_hom: calculates FHOM and reports basic statistics
}
