% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_roh.R
\name{F_roh}
\alias{F_roh}
\title{F_roh
F_roh() calculates FROH and returns basic statistics on this coefficient}
\usage{
F_roh(data_path, coverage)
}
\arguments{
\item{data_path}{A character string indicating the path of the .hom.indiv
PLINK output}

\item{coverage}{A character string indicating the path of the .bim file,
used to calculate the genome coverage. It is assumed that you have 22 chromosome in this case.
If the genome size is known it is possible to enter it as a numeric value}
}
\value{
@return A list, containing:
\item{FROH_values}{A data frame with FID, IID, and Froh values}
\item{Froh_summary_table}{A data frame with overall Froh summary statistics (Maximum, Minimum, Mean, SD)}
\item{by_population_Froh_stats}{A data frame with per population Froh summary statistics (Maximum, Minimum, Mean, SD)}
}
\description{
F_roh
F_roh() calculates FROH and returns basic statistics on this coefficient
}
