% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homozyg_density.R
\name{homozyg_density}
\alias{homozyg_density}
\title{homozyg_density}
\usage{
homozyg_density(bim_path)
}
\arguments{
\item{bim_path}{a PLINK .bim file}
}
\value{
snp_density
}
\description{
homozyg_density() determines the minimum SNP density (in Mb) for the segment
to be considered ROH
}
