% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLINK_pca.R
\name{PLINK_pca}
\alias{PLINK_pca}
\title{PLINK_pca
PLINK_pca() generates a PCA plot of the populations from a PLINK bed file}
\usage{
PLINK_pca(bed_file, pop)
}
\arguments{
\item{bed_file}{A character string indicating the path of the .bed file}

\item{pop}{A character string indicating the path of a text file with the
population names (identical to the first column in a .fam file)}
}
\value{
pca plot
}
\description{
PLINK_pca
PLINK_pca() generates a PCA plot of the populations from a PLINK bed file
}
