% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLINK_plotPCA.R
\name{PLINK_plotPCA}
\alias{PLINK_plotPCA}
\title{PLINK_plotPCA
PLINK_plotPCA() generates a PCA plot of the populations from a PLINK bed file}
\usage{
PLINK_plotPCA(bed_file, pop)
}
\arguments{
\item{bed_file}{A character string indicating the path of the .bed file}

\item{pop}{A character string indicating the path of a text file with the
population names (identical to the first column in a .fam file)}
}
\value{
pca_plot: ggplot object representing the PCA plot
}
\description{
PLINK_plotPCA
PLINK_plotPCA() generates a PCA plot of the populations from a PLINK bed file
}
\examples{
\dontrun{
PLINK_pca("data/genotype.bed","data/population.txt")
}

}
