% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relatedness_hm.R
\name{relatedness_hm}
\alias{relatedness_hm}
\title{relatedness_hm}
\usage{
relatedness_hm(data_path, output_dir = "figures/")
}
\arguments{
\item{data_path}{A character strings indicating the paths to the .genome file}

\item{output_dir}{A character strings indicating the output directory name.
If not indicated it will be "figures/" by default}
}
\value{
A heatmap for cryptic relatedness
}
\description{
relatedness_hm
}
