% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/functions_statictical_analysis_using_plink_outputs.R
\name{roh_stat}
\alias{roh_stat}
\title{This function performs basic statistics on the number and the length of ROH}
\usage{
roh_stat(indiv_path)
}
\arguments{
\item{data_path}{A character string indicating the path of the .hom.indiv PLINK output}
}
\value{
1 dataframe with elementary statistics
}
\description{
This function performs basic statistics on the number and the length of ROH
}
