% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatter_plot.R
\name{scatter_plot}
\alias{scatter_plot}
\title{scatter_plot
scatter_plot() processes multiple input files, associates data with specified
groups, generates individual scatter plots for each group, saves them in
separate files, and creates a combined grid plot for all groups}
\usage{
scatter_plot(file_paths, groups, output_dir = "figures/")
}
\arguments{
\item{file_paths}{A list of character strings indicating the paths to the
.hom.indiv files of each group}

\item{groups}{A character vector of all the groups}

\item{output_dir}{A character string with the name of the directory where the
plot will be saved. It is by default called "figures/"}
}
\value{
individual plots for each group and a combined plot of all the groups
}
\description{
scatter_plot
scatter_plot() processes multiple input files, associates data with specified
groups, generates individual scatter plots for each group, saves them in
separate files, and creates a combined grid plot for all groups
}
