% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/functions_statictical_analysis_using_plink_outputs.R
\name{F_hom}
\alias{F_hom}
\title{This function calculates FHOM and give basic statistics}
\usage{
F_hom(het_path)
}
\arguments{
\item{het_path}{A character string indicating the path of the .het PLINK output}
}
\value{
2 dataframe; the 1st having FHOM values and the 2nd having elementary statistics of FHOM
}
\description{
This function calculates FHOM and give basic statistics
}
