% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/functions_statictical_analysis_using_plink_outputs.R
\name{F_roh}
\alias{F_roh}
\title{This function calculates FROH and give basic statistics}
\usage{
F_roh(data_path, bim_path)
}
\arguments{
\item{data_path}{A character string indicating the path of the .hom.indiv PLINK output}

\item{bim_path}{A character string indicating the path of the .bim file, it is used to calculate the genome coverage,
if the genome size is known it is possible to enter it as a numeric value}
}
\value{
2 dataframe; the 1st having FROH values and the 2nd having elementary statistics of FROH
}
\description{
This function calculates FROH and give basic statistics
}
