% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_roh.R
\name{F_roh}
\alias{F_roh}
\title{F_roh
F_roh() calculates FROH and returns basic statistics on this coefficient}
\usage{
F_roh(data_path, coverage)
}
\arguments{
\item{data_path}{A character string indicating the path of the .hom.indiv
PLINK output}

\item{coverage}{A character string indicating the path of the .bim file,
used to calculate the genome coverage. If the genome size is known
it is possible to enter it as a numeric value}
}
\value{
2 dataframe. The 1st having FROH values and the 2nd having elementary
statistics of FROH
}
\description{
F_roh
F_roh() calculates FROH and returns basic statistics on this coefficient
}
