% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROH_identification_param_functions.R
\name{homozyg_snp}
\alias{homozyg_snp}
\title{homozyg_snp}
\usage{
homozyg_snp(hardy_path, ni)
}
\arguments{
\item{hardy_path}{a .hwe file generated by the --hardy function in PLINK}

\item{ni}{the number of individuals}
}
\value{
l
}
\description{
A function that calculates the minimum number of SNPs to have less than 5\% of ROH generated by chance (l) (Lencz et al., 2007)
}
