% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roh_stat.R
\name{roh_stat}
\alias{roh_stat}
\title{roh_stat
roh_stat() performs basic statistics on the number and the length of ROH}
\usage{
roh_stat(indiv_path)
}
\arguments{
\item{indiv_path}{A character string indicating the path of the .hom.indiv
PLINK output}
}
\value{
1 dataframe with elementary statistics
}
\description{
roh_stat
roh_stat() performs basic statistics on the number and the length of ROH
}
\examples{
# Create a toy dataset
toy_data <- data.frame(
  KB = c(1500, 2500, 4000, 1000, 3000),
  NSEG = c(2, 4, 6, 1, 3)
)
# Write the dataset to a temporary file
toy_file <- tempfile(fileext = ".txt")
write.table(toy_data, file = toy_file, sep = "\t", row.names = FALSE, col.names = TRUE, quote = FALSE)

# Run roh_stat()
roh_stat(toy_file)

}
