% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roh_stat.R
\name{roh_stat}
\alias{roh_stat}
\title{roh_stat
roh_stat() performs basic statistics on the number and the length of ROH}
\usage{
roh_stat(indiv_path)
}
\arguments{
\item{indiv_path}{A character string indicating the path of the .hom.indiv
PLINK output}
}
\value{
1 dataframe with elementary statistics
}
\description{
roh_stat
roh_stat() performs basic statistics on the number and the length of ROH
}
